refactor:  bring modules to the same form
