+++ src/album/album.controller.ts
@@ -15,22 +15,22 @@ import {
 import { StatusCodes } from 'http-status-codes';
 import Album from 'src/types/album';
 
-import { AlbumsService } from './albums.service';
+import { AlbumService } from './album.service';
 import { CreateAlbumDto } from './dtos/createAlbumDto';
 import { UpdateAlbumDto } from './dtos/updateAlbumDto';
 
 @Controller('album')
-export class AlbumsController {
-  constructor(private albumsService: AlbumsService) {}
+export class AlbumController {
+  constructor(private albumService: AlbumService) {}
 
   @Get()
   getAlbums() {
-    return this.albumsService.getAlbums();
+    return this.albumService.getAlbums();
   }
 
   @Get(':id')
   getAlbumById(@Param('id', new ParseUUIDPipe()) id: string) {
-    const album = this.albumsService.getAlbumById(id);
+    const album = this.albumService.getAlbumById(id);
     if (!album) {
       throw new HttpException(
         'Album with this id does not exist',
@@ -43,13 +43,13 @@ export class AlbumsController {
   @Post()
   @UsePipes(new ValidationPipe())
   createAlbum(@Body() createAlbumDto: CreateAlbumDto): Album {
-    return this.albumsService.createAlbum(createAlbumDto);
+    return this.albumService.createAlbum(createAlbumDto);
   }
 
   @Delete(':id')
   @HttpCode(204)
   deleteAlbumById(@Param('id', new ParseUUIDPipe()) id: string) {
-    const index = this.albumsService.deleteAlbumById(id);
+    const index = this.albumService.deleteAlbumById(id);
     if (index === -1) {
       throw new HttpException(
         'Album with this id does not exist',
@@ -64,7 +64,7 @@ export class AlbumsController {
     @Param('id', new ParseUUIDPipe()) id: string,
     @Body() updateAlbumDto: UpdateAlbumDto,
   ) {
-    const { index, updatedAlbum } = this.albumsService.updateAlbum(
+    const { index, updatedAlbum } = this.albumService.updateAlbum(
       id,
       updateAlbumDto,
     );
