+++ src/album/album.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { AlbumController } from './album.controller';
+import { AlbumService } from './album.service';
+
+@Module({
+  controllers: [AlbumController],
+  providers: [AlbumService],
+  exports: [AlbumService],
+})
+export class AlbumModule {}
