--- src/albums/albums.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AlbumsController } from './albums.controller';
-import { AlbumsService } from './albums.service';
-
-@Module({
-  controllers: [AlbumsController],
-  providers: [AlbumsService],
-  exports: [AlbumsService],
-})
-export class AlbumsModule {}
