+++ src/app.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UsersModule } from './users/users.module';
-import { ArtistsModule } from './artists/artists.module';
-import { AlbumsModule } from './albums/albums.module';
+import { UserModule } from './user/user.module';
+import { ArtistModule } from './artist/artist.module';
+import { AlbumModule } from './album/album.module';
 
 @Module({
-  imports: [UsersModule, ArtistsModule, AlbumsModule],
+  imports: [UserModule, ArtistModule, AlbumModule],
   controllers: [AppController],
   providers: [AppService],
 })
