--- src/users/users.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserController } from './controllers/user/user.controller';
-import { UserService } from './services/user/user.service';
-
-@Module({
-  controllers: [UserController],
-  providers: [UserService],
-})
-export class UsersModule {}
